clang-tlb-import: move functions to handle metadata into separate file
