+++ tools/typelib-clang-tlb-importer/CMakeLists.txt
@@ -92,6 +92,7 @@ add_executable(typelib-clang-tlb-importer
     main.cpp
     TypelibBuilder.cpp
     NamingConversions.cpp
+    HandleMetaData.cpp
     IgnoredOrRenamedTypes.cpp
     )
